Fix Windows issue and shelve issue
- On Windows trying to open the temp files during `add` and `edit`
commands led to "Permission Denied" errors. This was because we were
trying to open the files in an editor while the file was still open in
our script
    - Fixed this by not using the `with` command, opening the named temp
      file with `delete=False` and making sure to close the file before
      it is opened in an editor
- In python 2 `shelve.open` can not be used with the `with` command
because the shelve class does not implement a `__exit__()` method that
is required by the `with` command. This is available in python 3 but not
in python 2
    - Fixed this by using `contextlib.closing` which wraps any object we
      want to use `with` with in with the appropriate `__exit__()`
      method so that `with` does not complain while exiting
    - Using the `with` command is useful because we wouldn't need to
      implement all the try except blocks to make sure that the file is
      cleaned up properly if an error occurs